Change the module export syntax and bump the version to 0.1.0
Bumped the version as it could be a breaking change.  The only metod directly exported is requireAsync as the actual export.  Other public methods are exported as methods of this property.  Updated the README to reflect these changes and added a CHANGES.md and ROADMAP.md file to track these and future changes.