Move config into its own file
Moved config to separate file this will aid in moving exported methods to own files. Upgraded the deps and devDeps.  Added catch to gulp task loader to not load task that fail on require (ie. do not load task and throw if it does not work on your node version).