Fix resolver paths bug, upgrade memoize and remove unnecessary modules
Paths finder function had a bug in it so no paths were being created from given roots - fixed. Upgraded the memoize module so there is only two access points memoize and memoizeRegExp, you just need to supply a type option to say if it is a promise or node-callback.  This could be further refactored to include RegExp.  Using memoize got rid of need for Triple_Map as you can now specify 3 cacheParams for the lookup.